Render missions using a table
