+++ src/components/Missionslist.js
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
-import PropTypes from 'prop-types';
 import './styles/Missions.css';
 import { useSelector, useDispatch } from 'react-redux';
 import { getMissions } from './features/missions/missionsSlice';
+import Item from './Item';
 
 const Missions = () => {
   const dispatch = useDispatch();
@@ -15,55 +15,34 @@ const Missions = () => {
 
   if (isLoading) {
     return (
-      <div>
+      <div className="loading">
         <h2>Loading...</h2>
       </div>
     );
   }
 
   return (
-    <section className="missions">
-      <Titles />
-      {missionItems.map((mission) => (
-        <Item
-          Mission={mission.mission_name}
-          Description={mission.description}
-          key={mission.mission_id}
-        />
-      ))}
+    <section>
+      <table className="missions">
+        <thead>
+          <tr>
+            <th><h3>Missions</h3></th>
+            <th><h3>Description</h3></th>
+            <th><h3>Status</h3></th>
+            <th>Actions</th>
+          </tr>
+        </thead>
+        {missionItems.map((mission) => (
+          <Item
+            Mission={mission.mission_name}
+            Description={mission.description}
+            key={mission.mission_id}
+          />
+        ))}
+      </table>
       ;
     </section>
   );
-};
-
-const Titles = () => (
-  <div className="header">
-    <h3 className="name name-header">Missions</h3>
-    <h3 className="name name-header">Descriptions</h3>
-    <h3 className="name name-header">Status</h3>
-  </div>
-);
-
-const Item = (props) => {
-  const {
-    Mission, Description,
-  } = props;
-
-  return (
-    <section>
-      <div className="item-details">
-        <h3 className="name">{Mission}</h3>
-        <p className="description">{Description}</p>
-        <button type="button">Activate Member</button>
-        <button type="button">Join Mission</button>
-      </div>
-    </section>
-  );
-};
-
-Item.propTypes = {
-  Mission: PropTypes.string.isRequired,
-  Description: PropTypes.string.isRequired,
 };
 
 export default Missions;
