+++ src/components/styles/Item.css
@@ -1,0 +1,4 @@
+.loading {
+  display: flex;
+  justify-content: center;
+}
